--- src/by/it/NakovMaks/jd01_01/Main.java
@@ -1,8 +1,0 @@
-package by.it.NakovMaks.jd01_01;
-
-public class Main {
-
-    public static void main(String[] args) {
-
-    }
-}
