--- src/by/it/NakovMaks/jd01_01/TaskB1.java
@@ -1,33 +1,0 @@
-package by.it.NakovMaks.jd01_01;
-
-/*
-Выводим квадрат числа
-Напишите программу, которая считывает с клавиатуры целое число a и выводит квадрат этого числа (a * a).
-Для считывания данных с клавиатуры используйте метод nextInt() объекта класса Scanner.
-
-Создать Scanner можно так:
-Scanner sc=new Scanner(System.in);
-
-Прочитать число можно так:
-int i=sc.nextInt();
-
-Требования:
-1. Программа должна выводить текст.
-2. В программе необходимо создать объект типа Scanner.
-3. Программа должна считывать число типа Int с клавиатуры.
-4. Программа должна выводить квадрат считанного числа.
-
- */
-
-import java.util.Scanner;
-
-class TaskB1 {
-    public static void main(String[] args) {
-        System.out.println("Enter your number: ");
-        Scanner sc = new Scanner(System.in);
-        int enter = Integer.parseInt(sc.nextLine());
-        int enterX2 = enter * enter;
-        System.out.println("The result is " + enterX2);
-
-    }
-}
