--- src/by/it/NakovMaks/jd01_01/TaskB3.java
@@ -1,24 +1,0 @@
-package by.it.NakovMaks.jd01_01;
-/*
-        Уберите комментарии с части кода, чтобы на экран вывелось сообщение «C началом курса».
-
-        Требования:
-        1. Программа должна выводить на экран текст.
-        2. Нужно убрать комментарии в некоторых строках, при этом не менять остальные.
-        3. В программу нельзя добавлять новые строки или удалять имеющиеся.
-        4. Выведенный текст должен соответствовать заданию.
-*/        
-class TaskB3 {
-        public static void main(String[] args) {
-//        String s = "C";
-//        System.out.println("началом курса");
-//        System.out.println("C началом");
-//        System.out.println("курса");
-        System.out.print("C началом");
-//        System.out.println("курса");
-//        System.out.println(s);
-//        System.out.print("C началом!");
-        System.out.print(" ");
-        System.out.println("курса");
-    }
-}
