+++ src/by/it/nakov/jd01_01/Main.java
@@ -1,0 +1,8 @@
+package by.it.nakov.jd01_01;
+
+public class Main {
+
+    public static void main(String[] args) {
+
+    }
+}
