+++ src/by/it/nakov/jd01_01/TaskA1.java
@@ -1,0 +1,15 @@
+package by.it.nakov.jd01_01;
+/*
+напишите программу которая выводит на экран
+Hello world!
+
+(строка должна быть одна, ошибки в тексте не допускаются)
+Для ручной проверки запустите программу Ctrl+Shift+F10
+Для автоматической проверки откройте и запустите класс Test_jd01_01
+*/
+
+class TaskA1 {
+    public static void main(String[] args) {
+        System.out.println("Hello world!");
+    }
+}
